Move condition test
