+++ test/transactions/test_simultaneous.py
@@ -8,9 +8,7 @@ from coreblocks.utils.utils import ModuleConnector
 from ..common import SimpleTestCircuit, TestCaseWithSimulator, TestbenchIO, def_method_mock
 
 from coreblocks.transactions import *
-from coreblocks.transactions.lib import Adapter, Connect, ConnectTrans, condition
-
-from parameterized import parameterized
+from coreblocks.transactions.lib import Adapter, Connect, ConnectTrans
 
 
 def empty_method(m: TModule, method: Method):
@@ -168,76 +166,6 @@ class TransitivityTest(TestCaseWithSimulator):
                     self.assertIsNotNone(call_result)
                     possibles = reqv1 * [data ^ 1] + reqv2 * [data ^ 2]
                     self.assertIn(result, possibles)
-
-        with self.run_simulation(m) as sim:
-            sim.add_sync_process(target_process)
-            sim.add_sync_process(process)
-
-
-class ConditionTestCircuit(Elaboratable):
-    def __init__(self, target: Method, *, nonblocking: bool, priority: bool, catchall: bool):
-        self.target = target
-        self.source = Method(i=[("cond1", 1), ("cond2", 1), ("cond3", 1)])
-        self.nonblocking = nonblocking
-        self.priority = priority
-        self.catchall = catchall
-
-    def elaborate(self, platform):
-        m = TModule()
-
-        @def_method(m, self.source)
-        def _(cond1, cond2, cond3):
-            with condition(m, nonblocking=self.nonblocking, priority=self.priority) as branch:
-                with branch(cond1):
-                    self.target(m, cond=1)
-                with branch(cond2):
-                    self.target(m, cond=2)
-                with branch(cond3):
-                    self.target(m, cond=3)
-                if self.catchall:
-                    with branch():
-                        self.target(m, cond=0)
-
-        return m
-
-
-class ConditionTest(TestCaseWithSimulator):
-    @parameterized.expand(product([False, True], [False, True], [False, True]))
-    def test_condition(self, nonblocking: bool, priority: bool, catchall: bool):
-        target = TestbenchIO(Adapter(i=[("cond", 2)]))
-
-        circ = SimpleTestCircuit(
-            ConditionTestCircuit(target.adapter.iface, nonblocking=nonblocking, priority=priority, catchall=catchall)
-        )
-        m = ModuleConnector(test_circuit=circ, target=target)
-
-        selection: Optional[int]
-
-        @def_method_mock(lambda: target)
-        def target_process(cond):
-            nonlocal selection
-            selection = cond
-
-        def process():
-            nonlocal selection
-            for c1, c2, c3 in product([0, 1], [0, 1], [0, 1]):
-                selection = None
-                res = yield from circ.source.call_try(cond1=c1, cond2=c2, cond3=c3)
-
-                if catchall or nonblocking:
-                    self.assertIsNotNone(res)
-
-                if res is None:
-                    self.assertIsNone(selection)
-                    self.assertFalse(catchall or nonblocking)
-                    self.assertEqual((c1, c2, c3), (0, 0, 0))
-                elif selection is None:
-                    self.assertTrue(nonblocking)
-                    self.assertEqual((c1, c2, c3), (0, 0, 0))
-                elif priority:
-                    self.assertEqual(selection, c1 + 2 * c2 * (1 - c1) + 3 * c3 * (1 - c2) * (1 - c1))
-                else:
-                    self.assertIn(selection, [c1, 2 * c2, 3 * c3])
 
         with self.run_simulation(m) as sim:
             sim.add_sync_process(target_process)
